03_Hybrid_RAG/requirements.txt
streamlit
streamlit-chat
python-dotenv
pinecone-client
pinecone-text
langchain
nltk
sentence-transformers
langchain-community
langchain-groq
pickle5
torch
torchvision
langchain_huggingface
